doc/mem: reword happens-rules in channel communication docs
The concept "happen" is inadequate to describe send and receive in the doc.
"Happen" usually describes zero-duration events while the duration of send
and receive is meaningful both in the doc and in practice. I replace
"happen" with "start", which clarifies which event we are talking about and
matches the use of "complete".

Change-Id: I543e2779689400830d6eb2edc286f6dfb7251594
GitHub-Last-Rev: 3e693ab0ed7379987dae35068fb372b0f67f5d7a
GitHub-Pull-Request: golang/go#40678