database/sql: make Rows.Scan properly wrap underlying errors
The existing implementation uses fmt.Errorf() with the %v verb.
When an error was returned from an implementor of the Scanner interface,
it was not possible to check for such error using
errors.Is() and errors.As().
By using the %w verb, this error checking becomes possible.
It is expected to remain backward compatible with string based checks.
An unit test has been added to prove both "old" and "new" style tests
work as expected.

Fixes #38099

Change-Id: Iea667041dd8081d961246f77f2542330417292dc