go/types: decouple the type checker from go/ast
WIP: this is an experiment allowing the typechecker to be run against
multiple AST implementations, by replacing the concrete AST with a
collection of interfaces, and implementing these interfaces using
minimal-allocation wrappers arount the go/ast

Some hopefully superficial shortcuts were taken along the way. In
particular, the 'token' package was not entirely abstracted away.

All tests pass, and benchmarks show that this makes the typechecking
pass between 0 and 10% slower, depending on the package.

To do:
 - More benchmarking and optimization.
 - Move the guts of the typechecker to internal/types.
 - Minor inline shuffling of the syntax tree (e.g. transposing specs
   into decls).
 - Resolve the token/position abstraction.
 - Generate wrappers programmatically.
 - Abstract away the type constructors.

DO NOT SUBMIT
DO NOT REVIEW

Change-Id: I55dc0a77b591f404e831ac91040d0c32a0de8929