go/types: decouple the type checker from go/ast
WIP: this is an experiment allowing the typechecker to be run against
multiple AST implementations. This is done by replacing the concrete AST
with a collection of interfaces, and implementing these interfaces with
minimal-allocation wrappers around the go/ast AST.

Some hopefully superficial shortcuts were taken along the way. In
particular, the 'token' package was not entirely abstracted away.

All tests pass, and benchmarks show that this makes the typechecking
pass 35-50% slower with an additional 25% allocation. Notably types.Info
remains unchanged, so observed impact on heap usage is probably not so
significant.

Still to do, if we were to prepare this for revie:
 - More benchmarking.
 - Move the guts of the typechecker to internal/types.
 - Fix the token/position abstraction.
 - Generate all the wrappers programmatically.
 - Experiment with different access patterns that may improve
   performance (for example "reifying" wrappers earlier in certain
   places to eliminate unnecessary type switches and allocations).

DO NOT SUBMIT
DO NOT REVIEW

Change-Id: I55dc0a77b591f404e831ac91040d0c32a0de8929