cmd/compile: avoid generating duplicates, also handle interfaces and complex
Interfaces (and other builtin types) are handled differently
because this phase precedes lowering of builtins.  That means
that aggregate values can flow into phis, which requires a more
wholesale rewrite than the one done here, so instead, just
take the SelectN apart into pieces, and reassemble with a *Make
for later transformation.

Change-Id: I2a0c32f20660bf8b99e2bcecd33545d97d2bd3c6