doc/mem: reword happens-rules in channel communication docs
The concept "happen" is inadequate to describe send and receive in the doc.
"Happen" usually describes zero-duration events while the duration of send
and receive is meaningful both in the doc and in practice. I replace
"happen" with "start", which clarifies which event we are talking about and
matches the use of "complete".

Change-Id: I543e2779689400830d6eb2edc286f6dfb7251594
GitHub-Last-Rev: 590c0e7cc5b481c0a2e378fdb8827a8e3af63f0d
GitHub-Pull-Request: golang/go#40678