cmd/link: add new -dumpsymsat linker option
DO NOT REVIEW
DO NOT SUBMIT

When debugging linker problems, it is common to instrument the linker
at various spots to dump out information on selected symbols, as a
means of discovering what's happening during the link. This patch
introduces a general-purpose symbol dumping debug utility, intended to
make it possible to grab debugging dumps without having to edit and
rebuild the linker, or keep patches lying around.

The -dumpsymsat flag takes advantage of the recently added linked
phase benchmarking infrastructure, meaning that you can request a
symbol dump at each point the benchmark code marks the start of a
linker phase.

Flag argument is of the form

     {aux}?{symtypes}{qualifier}?@{phases}

where

 {aux} is a clause of the form [c*] where 'c' is one of

  A => enables dumping of aux relationships
  I => enables dumping of relocations with tgt symbol names and indices
  O => enables dumping of outer/sub relationships
  P => enables dumping of Pkg for each sym
  R => enables dumping of relocations

 {symtypes} is a clause of the form symtype{,symtype}*

 with symtype specifying a sym.Kind, e.g. STEXT, SDATA, or
 the pseudo-type "all" indicate all symbol types

 {qualifier} is a clause of the form

 "text" indicating that only the contents of ctxt.Textp should be dumped
 "data" indicating that only the contents of ctxt.datap should be dumped

 {phases} is a clause of the form phase{,phase}*

 with phase specifying a linker phase (ex: "deadcode", "linksetup", "symtab")
 of the pseudo-phase "all" to dump at all phases.

 Examples:

   - dump text symbols (with relocs) and NOPTRBSS symbols at dodata + final:

       [R]STEXT,SNOPTRBSS@dodata,final

   - dump all symbols at pclntab:

	   all@pclntab

   - dump all ctxt.datap symbols at address:

	   all{data}@address

   - dump SELFROSECT symbols at all phases:

	   SELFROSECT@all

   - dump all symbols at deadcode with aux info and outer/sub info:

       [AS]all@deadcode

The symbol dump format looks roughly like:

    %S236: runtime.assertE2I2<1> t=STEXT sz=163 val=0
      + R0: R_TLS_LE  o=5 a=0 tgt=<nil>
      + R1: R_CALL    o=63 a=0 tgt=runtime.getitab
      + R2: R_CALL    o=154 a=0 tgt=runtime.morestack_noctxt

where 236 is the symbol index and '%' indicates reachability.

Change-Id: Ibf9353a6f4f6118c5593f2bbcff9bc247b8e702c