cmd/compile: mark type descriptor symbols content-addressable
DO NOT REVIEW
DO NOT SUBMIT

It doesn't work, as the linker looks up type descriptor symbols
by name in DWARF generation. It works fine if DWARF is disabled.

For now, just hash the symbol's name. The general content hash
algorithm doesn't work, as type descriptors are likely to have
cycles (e.g. ptrToThis field, or circular types). We could build
a structural hash that is specific for types, but that is quite
involved. It would need to understand types (which the obj
package currently doesn't need to know), and need many special
cases. So, do the simple thing for now.

Change-Id: I65f837c209900b07e0c043b5993ba46a541de66c