encoding/json: implement Is on all errors
Allows users to check:

      errors.Is(err, & UnmarshalTypeError{})
      errors.Is(err, & UnmarshalFieldError{})
      errors.Is(err, & InvalidUnmarshalError{})
      errors.Is(err, & UnsupportedValueError{})
      errors.Is(err, & MarshalerError{})

which is the recommended way of checking for kinds of errors.

SyntaxError was implemented on https://go-review.googlesource.com/c/go/+/253037, and will also add As and Unwrap implementation in future CLs.

PS: I can make smaller CLs if you'd like. My though was to split the CLs as one for Is, another for As and a final one for Unwrap. If you think split by error or something else is better let me know. :)

Change-Id: I1f8a503b8fdc0f3afdfe9669a91f3af8d960e028
GitHub-Last-Rev: 563a3d6c1ca96bc5ba49642799675ce1a58d5108
GitHub-Pull-Request: golang/go#41360