cmd/compile: adjust mid-stack inlining within small functions
DO NOT SUBMIT

If F calls G calls H, it may be that inlining H into G
results in G growing too large to inline into F.  If
G contains an allocation that would be exposed as non-
escaping if G *is* inlined, then inlining H into G can
result in a startling slowdown.  The heuristic here is
designed to make this less likely to occur; if G is
"small" and contains exactly one call site, then subtract
a little bit from the inlinining budget for H to help
ensure that if H can be inlined into G, then G can still
be inlined into F.

If G contains more than 1 call, it G isn't really small
for inlining purposes, and thus there is nothing to be
gained by reducing the budget for H.

In addition, the best cost for a "throw" is now lower than
it used to be.

The exact numbers chosen here are the result of extensive
benchmarking; if these turn out to cause significant
real-world regressions, we'll need to revisit them.

This is also designed to mitigate benchmarking problems
observed in a follow-on CL that uses the max of the
inlining costs for the two arms of an OIF, rather than
their sum.

Updates #30548.

Change-Id: I6eeb63dd568eb515efb40c24be63a68949525a02