go/minitypes: an experiment in managing a dual object graph
This is experiment #1 for managing a dual object graph between
internal/types and go/types. For the purpose of agility, pare down
go/types to the point where it can only typecheck structs.

In this experiment, the dual graph is managed by pinning 'references' to
the internal/types objects. It's not pretty, but avoids maintaining
external object lookups.

DO NOT SUBMIT
DO NOT REVIEW

Change-Id: I5e2209b4e0e6816a0234b25b7338eb74407ecdba