testing: reduce memory allocation in Helper()
Store the PC instead of the string name of the function, and defer
that conversion until we need it.

Helper() is still relatively expensive in CPU time (few hundred ns),
but memory allocation is now constant for a test rather than linear in
the number of times Helper() is called.

Change-Id: I6565feb491513815e1058637d086b0374fa94e19
GitHub-Last-Rev: d5def339797179738383463f3020d115a8eb89dc
GitHub-Pull-Request: golang/go#38834