cmd/go/internal/modfetch/codehost: forward the error from the underlying git call
The `loadRefs()` method is called in several places in this file.
It sets `refsErr` on error, but `refsErr` is not always checked and returned.
In a common case, an "unknown revision" error is thrown, which is not the correct error.

Change-Id: Id6ee9c36e705a1e4cb76b8f5f1364ff011679925
GitHub-Last-Rev: 9fd9b3391b23a28ffb30b1a4f37f50dde97e4cad
GitHub-Pull-Request: golang/go#41565