[dev.fuzz] internal/fuzz: coordinate fuzzing across workers
Package fuzz provides common fuzzing functionality for tests built
with "go test", for programs that use fuzzing functionality in the testing
package, and for "go test" itself.

CoordinateFuzzing is the main entry point for the coordinator.
"go test" now calls this function instead of running tests when
invoked with -fuzz.

Change-Id: I3901c6a993a9adb8a93733ae1838b86dd78c7036