cmd/asm,cmd/internal/obj/ppc64: follow up for nop change in previous release
This undoes the change that was made to the Optab in asm9.go
at the end of the last release to add entries for NOPs with
operands in CL 247842. That change highlighted the fact that
cmd/asm/internal/asm/testdata/ppc64.s was outdated and the
comments in that file were not doing anything. It was essentially
only verifying that the instructions in the file assembled
without error but did not verify that the encoded instruction
words were correct. To resolve these the following changes were
done:

- Remove entries from Optab in cmd/internal/obj/ppc64/asm9.go for
NOPs with operands, which should never happen on ppc64x and don't
assemble to anything anyway and never have.
- Create a single file ppc64.s which contains the encodings for
instructions taken from the previous ppc64enc.s, adding some
instructions that were in ppc64.s but not in ppc64enc.s
- Fix a bug in the assembly of addex which was encoding the
instruction word incorrectly and was discovered while testing
this change.

Change-Id: Iaada1563b137849ad195fa88f32ecc9ab3e1e95f