cmd/compile: add AMD64 parameter register defs, Arg ops, plumb to ssa.Config and regalloc
This is partial plumbing intended to show how I think it should all work
when activated.  There are stubs in regalloc.go and stackalloc.go, the
one in stackalloc.go may require thought, I feel like the current plan,
that expects a LocalSlot based on a gc.Node, might be problematic.

Still missing:
  code in gc/ssa.go or AuxCall to set up incoming params
  all the result lowering and plumbing
  code in gc/ssa.go to record offsets etc.

Change-Id: Ica8c55a4526793446189725a2bc3839124feb38f