cmd/compile: delay expansion of OpArg until expand_calls
As it says, delay expansipon of OpArg to the expand_calls phase,
to enable (eventually) interprocedural SSA optimizations, and
(sooner) change to a register ABI.

Includes a round of cleanup to function names and comments,
largely to match the expanded scope of the functions.

Two functions that performed overlapping things were combined
into a single function that is called twice.

For #40724.

Change-Id: Icbb78947eaa39f17f2c1210d5c2caef20abd6571